Don't use cookie for test target
Test target doesn't produce anything and nothing depends on it.
Also it's often useful to run tests multiple times in a row, so
having a cookie which remembers that the tests were already ran
and inhibits subsequent test runs is needless. It also eliminates
the need for retest target.

Approved by:	portmgr (mat)
Differential Revision:	D3875


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@401117 35697150-7ecd-e111-bb59-0022644237b5