devel/py-envisage: Framework for building applications with plug-ins from Enthought
Envisage is a Python-based framework for building extensible applications, that
is, applications whose functionality can be extended by adding "plug-ins".
Envisage provides a standard mechanism for features to be added to an
application, whether by the original developer or by someone else. In fact,
when you build an application using Envisage, the entire application consists
primarily of plug-ins. In this respect, it is similar to the Eclipse and
Netbeans frameworks for Java applications.
Each plug-in is able to:
 * Advertise where and how it can be extended (its "extension points").
 * Contribute extensions to the extension points offered by other plug-ins.
 * Create and share the objects that perform the real work of the application
   ("services").
The Envisage project provides the basic machinery of the Envisage framework.

WWW: http://www.github.com/enthought/envisage

PR:		204430
Submitted by:	vladimir.chukharev gmail .dot com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@401706 35697150-7ecd-e111-bb59-0022644237b5