Fix the ALSA=OFF build after r402530.
Both extrapatch-no_alsa and patch-git_d6927712 touch the same region of
CMakeLists.txt, and the latter was failing to apply when the ALSA option was
disabled.

Fix it by keeping patch-git_d6927712 (which fixes the CFLAGS values
regardless of the value of the ALSA option) and replacing extrapatch-no_alsa
with some post-patch/post-configure changes that achieve the same end
result.

PR:		204858
MFH:		2015Q4


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@402545 35697150-7ecd-e111-bb59-0022644237b5