Add upstream patch to fix the build with ALSA 1.1.0.
Some checks were too strict and assuming SND_LIB_MINOR was always 0. This
was making ALSA detection fail and preventing the plugins from being built
when the ALSA option was on.

Thanks to Stephen Hurd (shurd@) for pointing to the cause of the problem and
providing an initial version of the patch.

This does not need to be merged into 2015Q4 because it still contains ALSA
1.0.29.

PR:		204488


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@402657 35697150-7ecd-e111-bb59-0022644237b5