Depend on flex and bison from ports on 9.
According to portsmon, benchmarks/fio's build is currently broken on 9 after
the update to 2.2.11:

  lex: can't open lex.yy.c

with a verbose build it's possible to see that this is because base's flex
on 9 is too old and does not support "-o foo", it requires "-ofoo".

Depending on flex from ports to fix the build is not enough, as then the
yacc call fails:

  /usr/bin/yacc -o y.tab.c -l -d -b y exp/expression-parser.y
  yacc: e - line 68 of "exp/expression-parser.y", syntax error
  %parse-param { long long *result }
  ^
  Makefile:312: recipe for target 'y.tab.c' failed

Fix it by depending on both textproc/flex and devel/bison when the versions
in base are too old.

PR:		205022
Approved by:	bapt (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@403016 35697150-7ecd-e111-bb59-0022644237b5