Enable lpr printing agian.
While configure happily reported lpr is enabled and the module was
installed. The print code wasn't configured to use lpr. The reason
for this is that the configure of this code moved from the gtk/Makefile
to configure. Remove the now obsolete sed line.

Reported by:	dinoex@
Tested by:	dinoex@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@404597 35697150-7ecd-e111-bb59-0022644237b5