java/wildfly(80|81|82): Deprecate three EOL versions of wildfly
There are currently no less than 5 versions of wildfly: 80, 81, 82, 90,
and 100.  According to the developers, only the current release is really
supported although some security patches may get backported to the
previous release as necessary.  Since the current release is 10.0, only
the last version of 9 could potentially receive updates.

I'm setting the expiration for 80 and 81 for 1 FEB 2016.
I'm setting the expiration for 82 for 15 JUL 2016

That should get 82 in the new quarterly branch and then it's gone from
the trunk.  That should be reasonable to upgrade to 10.0 or later by
then.

PR:		205490
Approved by:	yerenkow (gmail) (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@405620 35697150-7ecd-e111-bb59-0022644237b5