archivers/py3-libarchive-c: Add Python 3.x version of port
Add a Python 3.x sub-port of py-libarchive-c, in order for a port of the
Debian diffoscope project to be created, which is Python 3 only. This
(py3-*) hack^W workaround ensures a py3x-* package can be created by
default.

 - Allow USES to be overriden in base port accordingly

While I'm here

 - Enable NO_ARCH

Requested by:	emaste, bapt (for diffoscope, reproducible builds)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@407585 35697150-7ecd-e111-bb59-0022644237b5