TagSoup - Just Keep On Truckin'
TagSoup is a SAX-compliant parser written in Java that, instead of parsing
well-formed or valid XML, parses HTML as it is found in the wild: poor,
nasty and brutish, though quite often far from short.  TagSoup is designed
for people who have to process this stuff using some semblance of a rational
application design.  By providing a SAX interface, it allows standard XML
tools to be applied to even the worst HTML.  TagSoup also includes
a command-line processor that reads HTML files and can generate either
clean HTML or well-formed XML that is a close approximation to XHTML.

WWW: http://home.ccil.org/~cowan/tagsoup


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@408700 35697150-7ecd-e111-bb59-0022644237b5