Fix net-snmp ipv6 option handling when other options are also used.
r381568 added CONFIGURE_ARGS+=--with-out-mib-modules="mibII/ipv6", but the
handling of other options can result in two different --with-out-mib-modules
arguments being passed to configure, and only the second one has
effect. That leads to build failures because it tries to build the old
mibII/ipv6 which is no longer supported on freebsd.

This change adds mibII/ipv6 to the list of excluded modules that gets turned
into the --with-out-mib-modules argument, so it plays nice with other option
handling.

PR:		199401
Approved by:	mat(mentor)
Differential Revision:	https://reviews.freebsd.org/D5329


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@409215 35697150-7ecd-e111-bb59-0022644237b5