graphics/copperspice: partially revert r403659
Copperspice was already marked BROKEN (everywhere).  Adding a second
broken was unnecessary and masks the originally reason.  I've been waiting
for the authors to fix the issues -- they are enthusiastic about
supporting BSD but paid work and lack of test machines have been pushing
back the new work.  This effort is not dead though.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@409927 35697150-7ecd-e111-bb59-0022644237b5