lang/pypy: unbreak build when pypy isn't installed.
When pypy isn't installed then the port needs to use python to translate
pypy, however all options are not specified since two of the three (the
pypy related options) are not valid.  Overcome this by a bit of re-ordering
and abusing OPTIONS_SLAVE in the later case.

Reported by:	Andrzej Tobola <ato@iem.pw.edu.pl>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@411039 35697150-7ecd-e111-bb59-0022644237b5