Update CMake to 3.5.0.
Announcement: https://blog.kitware.com/cmake-3-5-0-available-for-download/

Most ports have already been fixed for this update, but science/orthanc was
included in this commit because I am not entirely sure the changes in it work
with CMake < 3.5.0. In short, FindDCMTK.cmake was rewritten and the way orthanc
uses the DCMTK variables was incompatible with the use upstream intended.

PR:		208033


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@411409 35697150-7ecd-e111-bb59-0022644237b5