redundant-opt-files script: Handle cache for deleted ports
The redundant-opt-file scripts assumes that if options are cached,
then the port that it was created by still exists.  This, of course, is
a terrible assumption.

If the cached options refer to a port that has since been deleted or
otherwise no longer exists, just print the origin and continue.  The
normal use case is that the origin will be piped to xargs rm -rf which
will purge the obsolete directory as desired.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@411809 35697150-7ecd-e111-bb59-0022644237b5