Restrict dns/ironsides to gcc5-aux only
Right now, ironsides emits a STORAGE ERROR during building with
lang/gcc6-aux.  It's unclear if the problem lies with compiler or with
ironsides.  For now, limit building it with gcc5-aux (the default).

The USES=ada:5 setting wasn't supported, but it should have been.  This
has also been fixed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@413032 35697150-7ecd-e111-bb59-0022644237b5