sysutils/mfid: Update to 0.02
The original implementation of mfid from Yahoo! disabled the RAID
controller's Patrol Reads. The user is given no warning that this is
happening. This update to mfid removes this functionality and comes with
a large pkg-message warning to inform the user about the previous
behavior as well as steps to re-enable Patrol Reads for their RAID card.

Thank you to Terrence Koeman for noticing this behavior, and jhb for
confirming the history of this code.

While here, change the MASTER_SITE to my LOCAL distfiles location.

MFH:		2016Q2


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@414074 35697150-7ecd-e111-bb59-0022644237b5