- Add rubygem-pathutil 0.9.0
Pathutil tries to be a faster pure Ruby impelementation of Pathname. It arose
out of a need to fix basic problems with Pathname, such as suscepetibility to
join overrides, need for automatic encoding, and normalization (for stuff like
Jekyll) and the ability to do other safe-style operations in an encapsulated
format, like copying files and folders with symlinks but only if they originate
from the given root.

WWW: https://github.com/envygeeks/pathutil


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@414302 35697150-7ecd-e111-bb59-0022644237b5