ports-mgmt/synth: Stop using bundled adacurses & IFM
Now that the Ada Framework is based on GCC6 by default, it is no
longer required for Synth to bundle and build a local version of
devel/adacurses and misc/ini_file_manager.

Switch the port to normal BUILD_DEPENDS and remove the bundle
distfile information.  No bump is required because these libraries
are statically linked into Synth, so there's no practical difference
between older and newer packages.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@414334 35697150-7ecd-e111-bb59-0022644237b5