Add new port database/adabase
Thick database bindings for MySQL, PostgreSQL and SQLite written in Ada.

This is the first release of AdaBase, an abstraction library that provides
a consistent interface to multiple database servers.  Currently only one
driver for MySQL is provided, but additional ones for SQLite and
PostgreSQL are planned for the near term.  It's extensible, so support
for others such as Firebird, Oracle and MSSQL would be easily possible.

AdaBase offers unique features over similar frameworks.  For starters,
it's limited to database support rather than including many other unwanted
components in a "kitchen sink" fashion, and unneeded drivers can be
excluded from the library as desired.  It's got a developer and commerce
friend license (ICS), it comes with good documentation and working
examples, and the bindings are thick enough where database server backends
can be interchangeable.

AdaBase may seem familiar to some users as it was partially inspired by
PHP's PDO database framework and is a sequel of sorts to an earlier
project by the same author, Pascal Data Objects.

WWW: http://jrmarino.github.io/AdaBase/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@414838 35697150-7ecd-e111-bb59-0022644237b5