The BIOM file format (canonically pronounced biome) is designed to be a general-use format for representing biological sample by observation contingency tables. BIOM is a recognized standard for the Earth Microbiome Project and is a Genomics Standards Consortium supported project.
The BIOM format is designed for general use in broad areas of comparative
-omics. For example, in marker-gene surveys, the primary use of this format is
to represent OTU tables: the observations in this case are OTUs and the matrix
contains counts corresponding to the number of times each OTU is observed in
each sample. With respect to metagenome data, this format would be used to
represent metagenome tables: the observations in this case might correspond to
SEED subsystems, and the matrix would contain counts corresponding to the number
of times each subsystem is observed in each metagenome. Similarly, with respect
to genome data, this format may be used to represent a set of genomes: the
observations in this case again might correspond to SEED subsystems, and the
counts would correspond to the number of times each subsystem is observed in
each genome.

WWW: http://biom-format.org/

PR:		209193
Submitted by:	Joseph Mingrone


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@414865 35697150-7ecd-e111-bb59-0022644237b5