multimedia/plexmediaserver: Improve installation
The installation has been changed to be more amenable to the
master/slave setup between this port and the -plexpass version. The
newest plexpass release has some new files and removed others which was
conflicting with the shared do-install. Instead we will have a simpler
do-install and handle any additional needs in the pkg-plist.

This is not getting a PORTREVISION bump as it is a no-op. The same files
get installed and this version runs just fine. There are two binaries
that aren't getting stripped that were previously stripped in the last
version of this package, but this is of no concern. These files are
being removed in the next major Plex release anyway.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@416725 35697150-7ecd-e111-bb59-0022644237b5