New port: net-im/telepathy-qt5.
This is the Qt5 counterpart of net-im/telepathy-qt4. It was worked on by
tcberner and myself.

From pkg-descr:
This is a library for Qt-based Telepathy clients.

Telepathy is a D-Bus framework for unifying real time communication,
including instant messaging, voice calls and video calls. It abstracts
differences between protocols to provide a unified interface for
applications.

WWW: http://telepathy.freedesktop.org/doc/telepathy-qt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@418671 35697150-7ecd-e111-bb59-0022644237b5