Add a hack in pkg to skip running predeinstall for javavmwrapper-2.5
During upgrades/reinstall javavmwrapper 2.5 pre deinstall script deletes the pkg
temporary files preventing to finish the upgrade.

A fixed version of javavmwrapper will be made soon, this hack will remain in the
ports tree and not in pkg(8) itself. This hack is made to not bother users is
only affecting upgrade/reinstall phase not proper deinstall

PR:		210313


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@419360 35697150-7ecd-e111-bb59-0022644237b5