japanese/libtomoe-gtk: fix illegal placement of USES
USES must be defined before any inclusions like <options> or <pre>,
otherwise they can't get registered as run dependencies.  The presence
of USES=python after <pre> was trying to build+run dependency on
python conditionally based on the presence of an installed file.

Since the condition can't be moved up due to the LOCALBASE variable,
I've moved python up but changed it to USES+=python:build so that
python will never be registered as a run dependency, but it will be
available if required for building.  The alternative is register
python as a run dependency in every case because the port can't stay
how it is.  I picked the more conservative option I think.

Approved by:	just-fix-it


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@419596 35697150-7ecd-e111-bb59-0022644237b5