sysutils/gosa: Mark BROKEN due to unsatisfied magic quotes option
This version of gosa requires magic quotes to be turned on, but this
contraversial option was last available with PHP 5.3, which has long
since been removed from the tree.  A newer version of gosa, 2.7.4,
was released on 9 May 2012 which may or may not have the same magic
quotes requirement.

PR:	204846


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@420444 35697150-7ecd-e111-bb59-0022644237b5