mail/sendmail: fix stage-qa error
The strip command for sbin/sendmail fails due to the permissions of
the program installed in the stage directory.  Change the staged sendmail
mode to 755 so that the strip command succeeds.  The pkg-list specifies
the final mode to be 2555.

PR:		211353
Reported by:	dave (ci.com.au)
Final fix:	marino@
Appproved by:	maintainer timeout


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@420495 35697150-7ecd-e111-bb59-0022644237b5