ports-mgmt/synth: Upgrade version 1.50 => 1.51
This release changes the watchdog time limits from a fixed value to
one that is affected by the average load.  Using fixed values didn't
work well on extremely (and intentionally) loaded servers, e.g. an
overage load of 200 on an server with ncpu of 32.

The base times which were long to account for loading have been reduced,
and a multiplier of average-5 load / ncpu rounded to the nearest tenth
has been implemented (with 1.0 as the floor value).  For example, a 4-core
machine with an average 5-minute load of 6.0 will multiply the base
timeout by 1.5.  This approach worked well in an extreme test against the
entire ports tree.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@421445 35697150-7ecd-e111-bb59-0022644237b5