dns/ldns: Fix SSL library support, tidy options
The makefile for the ldns was very hard to follow due to preprocessor-
like hunks changing the logic flow even across targets.  As part of the
SSL fix, I felt compelled to convert much of the options logic to the
modern options framework to make it simpler to follow the makefile
logic (as a side benefit the port is a bit more tidy).

I also reworked the FreeBSD 9 + GOST option logic to set the new
IGNORE_FreeBSD_9 variable if GOST was requested on FreeBSD 9 with the
base SSL library.  I believe the previous logic using the WITH_OPENSSL_PORT
knob and CONFLICTS setting was incorrect.

Approvd by:	SSL blanket


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@421646 35697150-7ecd-e111-bb59-0022644237b5