devel/qtscriptgenerator : switch from GOOGLE_CODE to USE_GITHUB
In r422140 devel/qtscriptgenerator was marked BROKEN, as GOOGLE_CODE
has gone swimming with the fishes.

The project is available on github via qt-labs projects [1] which seems
reasonably official.

The only difference in the distfile created by this, is that the Copyright
holder changes from 'nokia' to 'digia'.

[1] https://github.com/qt-labs/qtscriptgenerator

Approved by:  	mat (mentor)
PR:		212789


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@422406 35697150-7ecd-e111-bb59-0022644237b5