[NEW PORT] devel/gitless: Version control system built on top of Git
Gitless is an experimental version control system built on top of Git. Many
people complain that Git is hard to use. We think the problem lies deeper
than the user interface, in the concepts underlying Git. Gitless is an
experiment to see what happens if you put a simple veneer on an app that
changes the underlying concepts. Because Gitless is implemented on top of Git
(could be considered what Git pros call a "porcelain" of Git), you can always
fall back on Git. And of course your coworkers you share a repository with
need never know that you're not a Git aficionado.

WWW: https://github.com/sdg-mit/gitless

Approved by:	amdmi3 (mentor)
Differential Revision:	D8159


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@423401 35697150-7ecd-e111-bb59-0022644237b5