www/aws(-demos): Upgrade version 17.0.0 => 17.0.1
Using a separately built templates_parser is just too much work. Instead,
use the framework's GITHUB support to pull template_parser from github
and install in the source tree (as designed).  This allows removal and
simplification of several patches.

Rework the -demos port to install several missing files and maintain
the directory tree so that web_elements are relatively in the intended
location.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@423748 35697150-7ecd-e111-bb59-0022644237b5