Introduce USES=varnish
You can now set the default version of Varnish with DEFAULT_VERSIONS to
build Varnish and its modules against your desired Varnish release
train.

The default version of Varnish throughout the ports tree will remain as
www/varnish4 until at least 2017Q1 when it will be changed in HEAD to be
www/varnish5. This should allow plenty of time for migration and for
stabilization of the changes in Varnish 5.x.

Submitted by:	vanilla


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@424707 35697150-7ecd-e111-bb59-0022644237b5