devel/py-twisted: Allow building with Python 3.x
Twisted is in the midst of a porting (to Python 3) effort [1][2] and
it should accordingly be easier for users to both test and use
Twisted on FreeBSD built with Python 3.x.

This change removes to Python 2.x limitation to the port and
enables multiple version (Python 2 and 3) concurrent installation.

While I'm here:

 - Switch to autoplist and remove entry removing *.c files.
 - Conditionally strip shared libraries depending on Python version.
 - Add test target.

[1] https://twistedmatrix.com/documents/current/core/howto/python3.html
[2] See Also: Unported modules list: WRKSRC/twisted/python/_setup.py

Approved by:	python (with hat)
MFH:		2016Q4


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@425940 35697150-7ecd-e111-bb59-0022644237b5