Make frame pointer tracing in cdrtools(-devel) work correctly
The cdrtools ports compile and run a small program ('avoffset') after
their configuration phase, which is supposed to figure out the stack
direction, frame pointer offset to main(), and so on.  This program
relies on the compiler not optimizing away certain constructs, so after
the import of clang 3.9.0 into base, it got stuck in an endless loop.

Fix it by marking a few functions as noinline, and making one particular
variable volatile.

Reported by:	antoine
Approved by:	marius (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@427613 35697150-7ecd-e111-bb59-0022644237b5