New port: devel/py-dill
Dill extends python's 'pickle' module for serializing and de-serializing
python objects to the majority of the built-in python types. Serialization
is the process of converting an object to a byte stream, and the inverse
of which is converting a byte stream back to on python object hierarchy.

WWW: https://pypi.python.org/pypi/dill

PR:		214677
Submitted by:	Yuri Victorovich <yuri@rawbw.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@429482 35697150-7ecd-e111-bb59-0022644237b5