Update to 4.4.11
Changelog:

Under certain circumstances, bash will evaluate arithmetic expressions as
part of reading an expression token even when evaluation is suppressed. This
happens while evaluating a conditional expression and skipping over the
failed branch of the expression.

There is a race condition in add_history() that can be triggered by a fatal
signal arriving between the time the history length is updated and the time
the history list update is completed. A later attempt to reference an
invalid history entry can cause a crash.

Depending on compiler optimizations and behavior, the `read' builtin may not
save partial input when a timeout occurs.

Subshells begun to run command and process substitutions may attempt to
set the terminal's process group to an incorrect value if they receive
a fatal signal.  This depends on the behavior of the process that starts
the shell.

MFH:		2017Q1


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@432161 35697150-7ecd-e111-bb59-0022644237b5