Disable portscout for math/opensolaris-libm
Interestingly, since I created a release tag in the same date format
used by the PORTVERSION, portscout thinks there is a new version
available (2006.01.31 => 2017.01.24).

Since the PORTVERSION is never going to change, disable portscout
entirely.  As the maintainer of the github repo, I'll know when a new
release is available anyway.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@432426 35697150-7ecd-e111-bb59-0022644237b5