Stop installing the .pc file with a different name.
Use the default one: graphics/libkexiv2 has not existed in the tree since 2013,
and the x11/kdelibs4 update from r436971 removed a patch to FindKexiv2.cmake we
used to look for libkexiv2-kde4.pc. In practice, this commit unbreaks ports
that depended on graphics/libkexiv2-kde4 after the kdelibs4 update I mentioned.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@437019 35697150-7ecd-e111-bb59-0022644237b5