New port: sysutils/solaar
Solaar is a device manager for Logitech's Unifying Receiver.  It is
able to pair/unpair devices to the receiver, and for most devices read
battery status.

It comes in two flavors, command-line and GUI.  Both are able to list
the devices paired to a Unifying Receiver, show detailed info for each
device, and also pair/unpair supported devices with the receiver.

WWW: https://pwr.github.io/Solaar/

Approved by:	mat (mentor)
Differential Revision:	https://reviews.freebsd.org/D10155


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@437119 35697150-7ecd-e111-bb59-0022644237b5