devel/py-dateutil: Remove use of dos2unix
The file installed as dateutil-zoneinfo.tar.gz is corrupted by the dos2unix
run on all of the files in the distribution. There is no need to run dos2unix
as the only files with CRLF is the NEWS file, a README and a Windows BAT file
(which makes sense to use CRLF).

Accordingly:

- Remove use of USES=dos2unix
- Bump PORTREVISION to rebuild package (containing corrupted TZ data file)

PR:		217264
Submitted by:	gordon
MFH:		2017Q2


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@439790 35697150-7ecd-e111-bb59-0022644237b5