devel/cdecl: upgrade to 3.1
**C++ new-style casts**
Added support for C++ new-style casts.

**Language-sensitive autocompletion**
Autocompletion is now language-sensitive in that it will not offer completions that are not valid in the current language.

**Context-sensitive autocompletion**
Autocompletion is now slightly more context-sensitive in that it will offer only command keywords for the first word in the input.

**Fixed command-line error printing**
Command-line input is now printed when there's an error.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@442043 35697150-7ecd-e111-bb59-0022644237b5