graphics/vapoursynth-fmtconv: add new port
Requested by:	Juho Kerola

Fmtconv is a format-conversion plug-in for the Vapoursynth video
processing engine.

It does:
* Resizing.
* Bitdepth conversion with dithering.
* Colorspace conversion (matrix, transfer characteristics and chromatic adaptation).

It supports:
* Pixel data types: 8-–12-, 14- and 16-bit integer, 32-bit float.
* Colorspaces: RGB, Y, YUV in 4:4:4, 4:2:2, 4:2:0, 4:1:1 and YCgCo with the same chroma subsampling factors.
* Progressive and interlaced content.

Fmtconv is focussed primarily on quality and exactness rather than
execution speed.  This does not mean it is slow or unoptimized, but
fmtconv is clearly not on par with the fastest equivalent 8-bit
filters.

https://github.com/EleonoreMizo/fmtconv


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@443665 35697150-7ecd-e111-bb59-0022644237b5