Specifying an invalid PYTHON_VERSION against its USES should be IGNORED.
The port's own USES may note that is only supports certain versions.  If it
is attempted to build an unsupported version there's no reason to even
try.  Rather than giving a WARNING, actually mark it IGNORE.

Currently this should only impact devel/py3-enum34 which does not support
the default python3 version of 3.6.

With hat:	portmgr


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@443792 35697150-7ecd-e111-bb59-0022644237b5