devel/codesearch-py: Python library for accessing Chromium CodeSearch
The codesearch Python library provides an interface for talking to the
Chromium CodeSearch backend.

The primary entry point into the library is the codesearch class. Various
message classes you are likely to encounter are defined in messages.py

In addition, the library also includes facilities for maintaining an
ephemeral or persistent cache in order to minimize generated network
traffic.

WWW: https://github.com/chromium/codesearch-py

Reviewed by:	koobs
Differential Revision:	https://reviews.freebsd.org/D11336


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@444523 35697150-7ecd-e111-bb59-0022644237b5