- Update to an unofficial 5.1-1 which supports VirtualBox 5.1
Sadly upstream is very silent and has not released a new version
which supports VirtualBox 5.1, supports PHP 7.1 or fixes bugs. In
the meantime the community has created patches which are
collected in the github projects bugtracker and pull requests.
This unofficial version cherry picks the useful stuff of it.

Highlights of unofficial phpvirtualbox 5.1-1:

* based on git commit 65ebced (latest commit from May 19, 2016)

* Patch to support Virtualbox 5.1
https://github.com/phpvirtualbox/phpvirtualbox/pull/22

* Fix argument count mismatches revealed by PHP 7.1
https://github.com/phpvirtualbox/phpvirtualbox/pull/24

* Make decryption of images/VMs work by not providing a password ID
https://github.com/phpvirtualbox/phpvirtualbox/pull/55

* Tried to make clearOnSuspend work
https://github.com/phpvirtualbox/phpvirtualbox/pull/57

PR:		ports/220171
Submitted By:	rozhuk.im
With hat:	vbox@ (kind of)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@446746 35697150-7ecd-e111-bb59-0022644237b5