A package-based, source code amalgamater for collapsing Python packages into a single module.
The big idea here is to glue most of the source files in a package or
subpackage together into a single module, called
__amalgam__.py. Combined with some hooks in the __init__.py, this
should dramatically reduce the number of files that are being searched
for inside of the package. This is critical in larger projects where
import times are the major startup time cost.

WWW: https://github.com/xonsh/amalgamate

PR:		220596


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@447104 35697150-7ecd-e111-bb59-0022644237b5