Fix starting and stopping munin-asyncd. [1]
Turns out creating a start_cmd function is not enough, you have to add a
variable, start_cmd and point it to the start_cmd function.
As for stopping it, it should work out of the box, but for the fact that
munin-asyncd uses setproctitle or something, and it confuses the hell
out of rc.subr.

While there, pet rclint.

PR:		221000 [1] (based on)
Submitted by:	marcel herrbischoff com
Sponsored by:	Absolight


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@447210 35697150-7ecd-e111-bb59-0022644237b5