- update to 2.10.10 - add modified VDA patch to FILESDIR [1]
Note:
The 2.11 port will be deprecated at the time postfix-2.11.x is no longer
supported by the postfix author (Wietse Venema), this will happen with
the release of postfix-3.3.x.

[1] It seems the VDA project os dead, there are no working patches
against actual postfix releases provided by the project.
This is and will be the last postfix port with VDA support, users
using the VDA patch should adopt dovecot-lda which has also quota
support.  As alternative VDA users can jump on the wagon and try to
become an active maintainer of the SF.net VDA project (or a fork)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@447911 35697150-7ecd-e111-bb59-0022644237b5