Add py-pyPEG2.
I am currently in the process of creating ports for various components
related to CloudABI, including ARPC (an RPC framework that supports file
descriptor passing) and Flower (a capability-based network backplane).

The former comes with a script called 'aprotoc' for creating C++
bindings for messages and RPC services. This script depends on Python's
pyPEG2 package to parse (a subset of) Protobuf's .proto grammar.

Reviewed by:	wg
Differential Revision:	https://reviews.freebsd.org/D12048


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@448083 35697150-7ecd-e111-bb59-0022644237b5