PR/222002 initially patched graphics/xpdf to update to xpdf 4.00. As xpdf 4.00 introduces a brand new GUI based on QT5 rather than Motif, xpdf4 is being added as a new port, repocopied from graphics/xpdf.
Subsequent to this commit graphics/xpdf will be renamed to xpdf3
and a master port of graphics/xpdf (similar in concept to
security/krb5, sysutils/cfengine and sysutils/syslog-ng) will be
created to allow users to seamlessly update. People concerned
with POLA will be encouraged to use the to be renamed graphics/xpdf3
port.

PR:		222002
Submitted by:	Andrew Romanenko <melanhit@gmail.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@449702 35697150-7ecd-e111-bb59-0022644237b5