Add p5-MooseX-Util 0.006
MooseX::Util is a utility module that handles all of the same functions that
Moose::Util handles. In fact, most of the functions exported by this package are
simply re-exports from Moose::Util, so you're recommended to read the
documentation of that module for a comprehensive view.

WWW: http://search.cpan.org/dist/MooseX-Util/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@450534 35697150-7ecd-e111-bb59-0022644237b5