Fix packaging when DOCS=off
The manpages are once again only built when DOCS=on, but now so is the
default neomuttrc. The manpages are fine to protect with %%PORTDOCS%%
(not great, but permissible), but the default neomuttrc needs to be there.
So, generate it manually after the build. This is kludgey, but it at
least gets the port packaging again.

Approved by:	maintainer (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@452161 35697150-7ecd-e111-bb59-0022644237b5