Update to Wine 2.20.  This includes the following changes:  - Improved event support in MSHTML.  - Interpolation modes in Direct3D.  - Improved metafile support in GdiPlus.  - Initial version of Kerberos5 Authentication Package.  - OLE clipboard cache improvements.  - Various bug fixes.
Remove files/patch-partial-reservation.  This should not be needed any
longer since on FreeBSD the main wine program is now loaded at a lower
address, making more room available between Wine and 0x80000000, so it
may be ok now if the area above 0x80000000 stays reserved like on other
platforms. [1]

Mark the STAGING option (and hence the wine-staging port as well) as
BROKEN; that patchset still is not available yet, and we cannot wait
forever.

Discussed with: 	tijl [1]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@453481 35697150-7ecd-e111-bb59-0022644237b5