Update the URL's for Neverwinter Nights (Original and Diamond editions) data and client binaries due to them being moved again.
Since the available version of the Bink Player has changed and no
archives of the old copy exist, the ability to play movies in the game
is now experimental if and only if a newer version of the nwmovies.so
hook is compiled on Linux i386 and used.  The binary for the hook is not
currently provided, hence, it is experimental.  I am currently working
with the maintainer of that hook to make a release to add to the port to
bring back movie support.

Regenerate distinfo files to add TIMESTAMP.

Reviewed by:	lifanov, mat
Approved by:	lifanov
Differential Revision:	https://reviews.freebsd.org/D12960


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@453788 35697150-7ecd-e111-bb59-0022644237b5