Remove `groff' from USES.
It was recently killed in the base system and consumers have to use
`textproc/groff' port now, which is a real PITA: it pulls in some
obscure dependencies like `print/psutils', `graphics/netpbm', but
most importantly, `ghostscript' (unconditionally), which is a huge
abomination of its own, pulling another round of dependencies, some
of which are as crazy as `graphics/svgalib' (sic).  Trying to reset
all offered knobs to `off' results in unbuildable package.

Pass NROFF as mere `echo' and be done with it.  Generated manpages
remain identical.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@454682 35697150-7ecd-e111-bb59-0022644237b5