- Add devel/py{,3}-asttokens, missing dependency for py{,3}-flake8-import-order
The asttokens module annotates Python abstract syntax trees (ASTs)
with the positions of tokens and text in the source code that
generated them.

It makes it possible for tools that work with logical AST nodes to
find the particular text that resulted in those nodes, for example
for automated refactoring or highlighting.

WWW: https://pypi.python.org/pypi/asttokens

Approved by:	portmgr (bdrewery)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@454942 35697150-7ecd-e111-bb59-0022644237b5