www/nginx: Update brotli module
- Switch to fork that uses new brotli ABI

The upstream brotli module hasn't been updated for over a year even
though it's been broken for over a year with later brotli releases.
There's a fork that does update which uses the new brotli abi.
This patch is based on the fork. Requires some patching to the config
file as it depends on an old upstream brotli snapshot to be in the src dir.

PR:		224000
Reported by:	brnrd
Approved by:	tcberner (mentor)
Differential Revision:	https://reviews.freebsd.org/D13319


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@455319 35697150-7ecd-e111-bb59-0022644237b5