add devel/py-backports - namespace shim for Python backports
Currently, all py-backports.* ports conflict with each other
over a file that has the same contents.

This creates a port that installs the shared file and py-backports.*
ports should depend on this one and remove the shared file from their
own pkg-plist. It's the same approach as the one taken by OpenBSD.

Some ports are already converted as an example:
- security/py-backports.ssl_match_hostname
- devel/py-backports.functools_lru_cache
- devel/py-backports.shutil_get_terminal_size

The last two conversions and the fix to work with FLAVORS were submitted
by antoine. Thank you!

Reviewed by:	antoine
Approved by:	portmgr (antoine)
Differential Revision:	https://reviews.freebsd.org/D11095


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@455585 35697150-7ecd-e111-bb59-0022644237b5