audio/sndio: Update to 1.4.0
There are no code changes in libsndio, but the major/minor of libsndio
was bumped to 7.0 in this release to match libsndio on OpenBSD
regardless.  On OpenBSD some private symbols were hidden via a version
script.  This is not yet part of the portable release.  Avoid causing
unnecessary churn and keep using 6.1 for the time being.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@458929 35697150-7ecd-e111-bb59-0022644237b5